+++ zanata-war/src/main/java/org/zanata/service/impl/TranslationUpdatedManager.java
@@ -1,19 +1,31 @@
 package org.zanata.service.impl;
 
 import javax.enterprise.context.RequestScoped;
-import javax.enterprise.event.Event;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import com.google.common.base.Optional;
 import org.apache.deltaspike.core.api.provider.BeanManagerProvider;
+import org.zanata.ApplicationConfiguration;
 import org.zanata.async.Async;
+import org.zanata.common.LocaleId;
+import org.zanata.dao.DocumentDAO;
+import org.zanata.dao.PersonDAO;
 import org.zanata.dao.TextFlowDAO;
 import org.zanata.events.DocumentStatisticUpdatedEvent;
 import org.zanata.events.TextFlowTargetStateEvent;
+import org.zanata.events.TranslationUpdatedEvent;
+import org.zanata.model.HDocument;
+import org.zanata.model.HPerson;
+import org.zanata.model.HProject;
+import org.zanata.model.WebHook;
+import org.zanata.rest.dto.User;
+import org.zanata.rest.editor.service.UserService;
 import org.zanata.service.TranslationStateCache;
 
 import com.google.common.annotations.VisibleForTesting;
 import lombok.extern.slf4j.Slf4j;
+import org.zanata.util.UrlUtil;
 
 import javax.enterprise.event.Observes;
 import javax.enterprise.event.TransactionPhase;
@@ -21,9 +33,9 @@ import javax.enterprise.event.TransactionPhase;
 /**
  * Manager that handles post update of translation. Important:
  * TextFlowTargetStateEvent IS NOT asynchronous, that is why
- * DocumentStatisticUpdatedEvent is used for webhook processes. See
- * {@link org.zanata.events.TextFlowTargetStateEvent} See
- * {@link org.zanata.events.DocumentStatisticUpdatedEvent}
+ * DocumentStatisticUpdatedEvent is used for webhook processes.
+ * See {@link org.zanata.events.TextFlowTargetStateEvent}
+ * See {@link org.zanata.events.DocumentStatisticUpdatedEvent}
  *
  * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
  */
@@ -38,7 +50,20 @@ public class TranslationUpdatedManager {
     @Inject
     private TextFlowDAO textFlowDAO;
 
+    @Inject
+    private DocumentDAO documentDAO;
 
+    @Inject
+    private PersonDAO personDAO;
+
+    @Inject
+    private UserService userService;
+
+    @Inject
+    private ApplicationConfiguration applicationConfiguration;
+
+    @Inject
+    private UrlUtil urlUtil;
 
     /**
      * This method contains all logic to be run immediately after a Text Flow
@@ -50,6 +75,8 @@ public class TranslationUpdatedManager {
             TextFlowTargetStateEvent event) {
         translationStateCacheImpl.textFlowStateUpdated(event);
         publishAsyncEvent(event);
+
+        processWebHookEvent(event);
     }
 
     // Fire asynchronous event
@@ -68,10 +95,56 @@ public class TranslationUpdatedManager {
         }
     }
 
+    void processWebHookEvent(TextFlowTargetStateEvent event) {
+        HPerson person = personDAO.findById(event.getActorId());
+        if(person == null) {
+            return;
+        }
+        HDocument document = documentDAO.findById(event.getDocumentId());
+        String docId = document.getDocId();
+        String versionSlug = document.getProjectIteration().getSlug();
+        HProject project = document.getProjectIteration().getProject();
+        String projectSlug = project.getSlug();
+
+        int wordCount = textFlowDAO.getWordCount(event.getTextFlowId());
+
+        User user = userService.transferToUser(person.getAccount(),
+            applicationConfiguration.isDisplayUserEmail());
+
+        String url = urlUtil
+            .fullEditorTransUnitUrl(projectSlug, versionSlug,
+                event.getLocaleId(),
+                LocaleId.EN_US, docId, event.getTextFlowId());
+
+        TranslationUpdatedEvent webhookEvent =
+            new TranslationUpdatedEvent(user, project.getSlug(),
+                versionSlug, docId, event.getLocaleId(), url,
+                event.getPreviousState(), event.getNewState(), wordCount);
+
+        for (WebHook webHook : project.getWebHooks()) {
+            publishWebhookEvent(webHook, webhookEvent);
+        }
+    }
+
+    void publishWebhookEvent(WebHook webHook,
+        TranslationUpdatedEvent translationUpdatedEvent) {
+        WebHooksPublisher.publish(webHook.getUrl(), translationUpdatedEvent,
+            Optional.fromNullable(webHook.getSecret()));
+        log.debug("firing webhook: {}:{}", webHook.getUrl(),
+            translationUpdatedEvent);
+    }
+
     @VisibleForTesting
     public void init(TranslationStateCache translationStateCacheImpl,
-            TextFlowDAO textFlowDAO) {
+        TextFlowDAO textFlowDAO, DocumentDAO documentDAO, PersonDAO personDAO,
+        UserService userService,
+        ApplicationConfiguration applicationConfiguration, UrlUtil urlUtil) {
         this.translationStateCacheImpl = translationStateCacheImpl;
         this.textFlowDAO = textFlowDAO;
+        this.documentDAO = documentDAO;
+        this.personDAO = personDAO;
+        this.userService = userService;
+        this.applicationConfiguration = applicationConfiguration;
+        this.urlUtil = urlUtil;
     }
 }